Time: 12 ms (25.77%) | Memory: 55.3 MB (36.97%) - LeetSync

+++ 1956-maximum-element-after-decreasing-and-rearranging/maximum-element-after-decreasing-and-rearranging.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     int maximumElementAfterDecrementingAndRearranging(vector<int>& arr) {
         sort(arr.begin(), arr.end());
-        long long i = 0;
-        long long n = arr.size();
+        int i = 0;
+        int n = arr.size();
 
             arr[0] = 1;
